netutils/libcoap : Fix problem from static analysis tool
This commit is patch for fixing problems fro static analysis tool
- Add check return value of pthread_attr_init

+++ apps/netutils/libcoap/pdu.c
@@ -994,7 +994,7 @@ coap_pdu_t *coap_convert_to_tcp_pdu(coap_pdu_t *pdu)
 		transport = coap_get_tcp_header_type_from_size((pdu->length - COAP_UDP_HEADER));
 	}
 
-	if ((pdu->length - COAP_UDP_HEADER) < 0) {
+	if ((int)(pdu->length - COAP_UDP_HEADER) < 0) {
 		newsize = 0;
 	} else {
 		newsize = (pdu->length - COAP_UDP_HEADER);
